Build system updates
Updated the Gulp build system to accomplish the following:

1. De-compartmentalized the build system code into a single
`gulpfile.babel.js`. Not only is this easier to maintain, but we are
now using modern ECMAscript to define tasks.
2. Remove numerous NPM scripts in favor of a handful of useful, succint
ones. Instead of having an NPM script for each and every variation, we
now have a handful that allows for easier development/deployment.
3. Tap into in the `NODE_ENV` environment variable to define
development/production assets. This is all handled through the NPM
scripts.
4. Get rid of unminified/minified files. While in development, assets
will be unminified. Once built, the assets will be minified accordingly.
This means that there's no `.min` files of assets anymore.
5. Handle image optimization/resizing. Images were previously unhandled
prior to this update, making local development a pain. Now, images can
be found in the source directory, and are automatically
optimized/resized by the build system if needed before placing them in
the `dist` directory.
6. Localhost development with live reloading. When you run `npm run
develop`, not only will the build system build out all development
assets, but it will spin up a local server and watch for changes. When
you make a change, you will instantly see them take effect in your local
server browser window.