SASS fixes and changes
Updated the home and project SASS to accomplish the following:

1. Remove the unecessary horizontal scroll-bar on the homepage by
changing the `#wrapper` behavior to only scroll in the y-direction.
2. Make the navigation buttons on mobile devices larger. I noticed while
browsing the site on my phone, the links in the bottom nav on both the
homepage and project pages were too small. This change should make them
easier tap-targets.
3. Made the buttons in the project page nav that reveal a section
instead of opening a link proper HTML `<button>` elements for semantic
validity and accessibility reasons.