Switch from NodeSass to DartSass
With the announcement from the Sass core team regarding the deprecation
of LibSass (which NodeSass is a wrapper for), it makes sense to switch
the Sass compiler used under the hood for the Gulp Sass task to use Dart
Sass. While this has not caused any issues in terms of compilation, it
opens the doors for a wider range of features to tap into, which will be
critical in the longevity of this project.