JS optimizations and changes
Updated the home and project JS to clean up the code, use modern syntax,
and simplify everything.

This change also introduced changes to the way sections are
hidden/shown. Previously, it used a `.hidden` and `.visible` class to
accomplish this. Now, it sections are hidden by default, and made
visible when their `aria-hidden` attribute is set to false.