Further restructuring and feature implementation
In order to foster better maintainability, the following restructuring
changes have been made:

1. All styles, functionality, and structure of a component can be found
in its own directory under the new `components` directory; This will
make it easier to identify and work on all of these aspects on a
per-component level.
2. All projects are now in their own dedicated `projects` directory at
the root.
3. Shared PUG files (for the layouts) are now under the `shared`
directory.
4. The homepage PUG and site-level SCSS/JS are at the root of the repo.
5. Instead of building everything into `dist/illustrative-code`, this is
now handled in the Gulpfile's `serve()` function; This will make it
easier to move the site to a new location or subdirectory in the future.
6. All nav/card component functionality have been reimplemented to
accomplish feature parity with older versions.
7. SCSS classes now entirely use BEM naming conventions.