Styles removed from project nav
With the addition of projects with JS, and the questionable need to show
production code when a link to the project's source code on Github, the
"Styles" link has been dropped from the nav.

In addition, to remove confusion between the homepage's "Github" nav
link and the project's link of the same name, the project link has been
renamed to "Code" and uses a `</>` icon instead.