Better SEO and PWA features
To bring Illustrative Code into the modern web, a few things had to be
done. This includes providing better metadata for search engines and
social platforms to display more useful content. In addition, providing
basic PWA features lays the foundation for more features in the future.

This change starts at the new `config.js`, which defines useful data
that permeates through the entire build system, allowing for easily
generating favicons, to quickly changing the site root and build
directory if need be in the future.

In addition, a new `service-worker.js` is registered on page load by the
`site.js`, allowing for powerful PWA features such as asset caching.

Also, a new `images/icon.png` is included to generate all new favicons,
Apple touch icons, and PWA icons with a single Gulp task. And a new
`images/thumbnail.png` is included for an `og:image` meta tag on the
homepage to showcase what the project is all about.