Offline page
Implemented an offline page to be displayed when a visitor has the site
cached and is offline. This page, if available, will display up to 5
cached project pages to provide better navigability.

In addition, the `service-worker.js` was updated not only to accomodate
the new offline page, but to convert all promise-based logic to
async/await for cleaner and more readable code. It also had its caches
renamed and no longer attempts to cache BrowserSync files when
developing locally.