Webpack config for HTML files
Updated the `webpack.config.js` file to have 2 exports, one for JS, and
the other for HTML. The HTML config is needed to convert page JS files
to CommonJS before running through the `jsToHTML(...)` Gulp plugin. It
also allows for the import of HTML, SVG as strings and MD converted to
an HTML string.