SVG icon changes
Following how Font Awesome's React library adds crucial attributes to
their SVG icons, I've accomplished something similar with an updated
`Icon(...)` function, which takes an SVG icon as an argument and applies
these crucial attributes to it.

In addition, the `logo.svg` has been updated to have the same attributes
asside from the class, which will continue to be `.logo`.