Nav link fix
When creating a nav link, if the `href` was to an internal page, it was
adding bogus `rel` and `target` attributes that were creating issues in
PWA installs (namely, on Windows, clicking the "Back" button from any
project would open the homepage in your preferred browser). This has
been fixed.