Dependency updates
Updated all outdated dependencies to their latest versions while bumping
`eslint` and `marked` to a new major version.

With the `marked` upgrade, the `webpack.config.js` had to be updated to
import the dependency with a named export instead of the default.

With the `eslint` upgrade, `babel-eslint` was removed and its RC file
was regenerated using more modern defaults.